Addition of Relay task and relay fixes

+++ Common/Inc/Relay.h
@@ -13,6 +13,6 @@ uint8_t Relay_IsOn();
 void Relay_TurnOn();
 void Relay_TurnOff();
 void Relay_SM();
-void Relay_SetThreshold(uint16_t threshold);
+void Relay_SetThreshold(uint32_t threshold);
 
 #endif
